vänförfrågningar funkar

+++ Datingsida/Controllers/FriendRequestModelController.cs
@@ -17,15 +17,13 @@ namespace Datingsida.Controllers
 {
     public class FriendRequestModelController : Controller
     {
-        private readonly DatingDbContext _context;        
+        private readonly DatingDbContext _context;
         private readonly UserManager<IdentityUser> _userManager;
-        private readonly ApplicationDbContext _application;
 
-        public FriendRequestModelController(DatingDbContext context, UserManager<IdentityUser> userManager, ApplicationDbContext application)
+        public FriendRequestModelController(DatingDbContext context, UserManager<IdentityUser> userManager)
         {
-            _context = context;            
+            _context = context;
             _userManager = userManager;
-            _application = application;
         }
         public ActionResult Index()
         {
@@ -34,10 +32,8 @@ namespace Datingsida.Controllers
 
         public ActionResult AddFriend(string userReceiver)
         {
-            //ProfileModel user = new ProfileModel();
-            //ProfileModel user1 = new ProfileModel();
             var friendrequest = new FriendList
-            {                
+            {
                 Status = false,
                 UserReceiver = userReceiver,
                 UserSender = _userManager.GetUserId(User)
@@ -45,51 +41,48 @@ namespace Datingsida.Controllers
 
             _context.Add(friendrequest);
             _context.SaveChanges();
-            return RedirectToAction("GetFriendRequests");
+            return RedirectToAction("Friends");
         }
 
-        public ActionResult AcceptFriendRequest(string UserSender, bool isAccepted)
+        public ActionResult AcceptFriendRequest(string UserSender, bool? isAccepted)
         {
             var user = _userManager.GetUserId(User); //Hämtar ID för den inloggade användaren
             if (ModelState.IsValid)
             {
-                //Hämtar dbc där den som tar emot friendreq är den inloggade användaren och någon annan skickar friendreq
-                var request = _context.Friendlists.FirstOrDefault(u => u.UserSender.Equals(UserSender) && u.UserReceiver.Equals(user));
-                request.Status = isAccepted;
-
-                if (request.Status)
-                {   
-                    //Kollar om request har ändrat state, gått från true till false eller vice versa
-                    _context.Entry(request).State = EntityState.Modified; 
-                }
-                else
+                if (isAccepted != null)
                 {
-                    var removeFriendRequest = _context.Request.Find(request.FriendRequestID);
-                    _context.Request.Remove(removeFriendRequest);
+                    //Hämtar dbc där den som tar emot friendreq är den inloggade användaren och någon annan skickar friendreq
+                    var request = _context.Friendlists.FirstOrDefault(u => u.UserSender.Equals(UserSender) && u.UserReceiver.Equals(user));
+                    request.Status = (bool)isAccepted;
+
+                    if (request.Status)
+                    {
+                        //Kollar om request har ändrat state, gått från true till false eller vice versa
+                        _context.Entry(request).State = EntityState.Modified;
+                    }
+                    else
+                    {
+                        var removeFriendRequest = _context.Request.Find(request.FriendRequestID);
+                        _context.Request.Remove(removeFriendRequest);
+                    }
+                    _context.SaveChanges();
+                    return RedirectToAction("Friends");
                 }
-                _context.SaveChanges();
-                return RedirectToAction("Friends");
             }
             return View();
         }
 
-        public ActionResult GetFriendRequests()
+        public string GetFriendRequests()
         {
             var user = _userManager.GetUserId(User);
             var count = _context.Friendlists.Count(u => u.UserReceiver.Equals(user) && !u.Status);
-            return RedirectToAction("Friends");
-
+            return count.ToString();
         }
 
         public ActionResult Friends()
         {
-            var friendlist = new FriendListViewModel();            
+            var friendlist = new FriendListViewModel();
             var user = _userManager.GetUserId(User);
-            
-            IEnumerable<FriendListViewModel> enumerableFriendProfiles;
-
-
-
             var requests = _context.Friendlists.Where(u => u.UserReceiver.Equals(user) || u.UserSender.Equals(user)).ToList();
 
             if (requests != null && requests.Any())
@@ -99,86 +92,70 @@ namespace Datingsida.Controllers
                     if (friendrequest.Status && friendrequest.UserReceiver.Equals(user))
                     {
                         var dbUser = _context.Profiles.Where(u => u.OwnerId == friendrequest.UserSender).First();
-                        friendlist.setFriends(dbUser);                        
-                        enumerableFriendProfiles = (IEnumerable<FriendListViewModel>)friendlist.friends;
-                        //var dbUser = _userManager.Users.Where(u => u.Id == sök.UserSender).First();
-                        //var datingdbuser = _context.Profiles.Where(u => u.Id == int.Parse(sök.UserSender)).First();
-                        //sök.UserSender = viewmodel.Id.ToString();
-                        //viewmodel.FirstName = datingdbuser.FirstName;
-                        //viewmodel.Age = datingdbuser.Age;
-                        //viewmodel.Presentation = datingdbuser.Presentation;
-                        //viewmodel.Gender = datingdbuser.Gender;
-                        //friendlist.setFriends(viewmodel);
-                        //friendlist.Status = sök.Status;
+                        friendlist.setFriends(dbUser);
+                        List<ProfileModel> lst = new List<ProfileModel>();
+                        lst = friendlist.friends;
+                        IEnumerable<ProfileModel> enumerableFriendProfiles = lst;
+
                         return View(enumerableFriendProfiles);
                     }
                     else if (friendrequest.Status && friendrequest.UserSender.Equals(user))
                     {
                         var dbUser = _context.Profiles.Where(u => u.OwnerId == friendrequest.UserReceiver).First();
                         friendlist.setFriends(dbUser);
-                        enumerableFriendProfiles = (IEnumerable<FriendListViewModel>)friendlist.friends;
-                        //var dbUser = _userManager.Users.Where(u => u.Id == sök.UserReceiver).First();
-                        //var datingdbuser = _context.Profiles.Where(u => u.Id == int.Parse(sök.UserReceiver)).First();
-                        //sök.UserReceiver = viewmodel.Id.ToString();
-                        //viewmodel.FirstName = datingdbuser.FirstName;
-                        //viewmodel.Age = datingdbuser.Age;
-                        //viewmodel.Presentation = datingdbuser.Presentation;
-                        //viewmodel.Gender = datingdbuser.Gender;
-                        //friendlist.setFriends(viewmodel);
-                        //friendlist.Status = sök.Status;
+                        List<ProfileModel> lst = new List<ProfileModel>();
+                        lst = friendlist.friends;
+                        IEnumerable<ProfileModel> enumerableFriendProfiles = lst;
+
                         return View(enumerableFriendProfiles);
                     }
                 }
-                return NotFound();
             }
-            return NotFound();
-
+            TempData["noFriend"] =  ViewData["noFriends"] = true;
+            TempData["noFriendRequest"] = ViewData["noFriendRequests"] = false;
+  
+            return RedirectToAction("NoFriendOrFriendRequest");
         }
 
-        public ActionResult ShowFriendRequests()
+        public ActionResult Requests()
         {
-            var friendlist = new FriendListViewModel();
-            var user = _userManager.GetUserId(User);  //inloggad användarID          
-            List<ProfileModel> allProfiles = _context.Profiles.ToList();
+            var friendRequests = new List<ProfileModel>();
+            IEnumerable<ProfileModel> enumerableFriendRequests = null;
 
-
-            var requests = _context.Friendlists.Where(u => u.UserReceiver.Equals(user) || u.UserSender.Equals(user)).ToList();
+            var user = _userManager.GetUserId(User);
+            var requests = _context.Friendlists.Where(u => u.UserReceiver == user).ToList();
 
             if (requests != null && requests.Any())
             {
-                foreach (var sök in requests)
+                foreach (var request in requests)
                 {
-                    if (sök.Status && sök.UserReceiver.Equals(user))
+                    if (!request.Status)
                     {
-
-                        //var dbUser = _userManager.Users.Where(u => u.Id == sök.UserSender).First();
-                        //var datingdbuser = _context.Profiles.Where(u => u.Id == int.Parse(sök.UserSender)).First();
-                        //sök.UserSender = viewmodel.Id.ToString();
-                        //viewmodel.FirstName = datingdbuser.FirstName;
-                        //viewmodel.Age = datingdbuser.Age;
-                        //viewmodel.Presentation = datingdbuser.Presentation;
-                        //viewmodel.Gender = datingdbuser.Gender;
-                        //friendlist.setFriends(viewmodel);
-                        //friendlist.Status = sök.Status;
-
-                    }
-                    else if (sök.Status && sök.UserSender.Equals(user))
-                    {
-                        //var dbUser = _userManager.Users.Where(u => u.Id == sök.UserReceiver).First();
-                        //var datingdbuser = _context.Profiles.Where(u => u.Id == int.Parse(sök.UserReceiver)).First();
-                        //sök.UserReceiver = viewmodel.Id.ToString();
-                        //viewmodel.FirstName = datingdbuser.FirstName;
-                        //viewmodel.Age = datingdbuser.Age;
-                        //viewmodel.Presentation = datingdbuser.Presentation;
-                        //viewmodel.Gender = datingdbuser.Gender;
-                        //friendlist.setFriends(viewmodel);
-                        //friendlist.Status = sök.Status;
+                        var senderProfile = _context.Profiles.Where(u => u.OwnerId == request.UserSender).First();
+                        // ProfileIndexViewModel users = new ProfileIndexViewModel();
+                        //UserViewModel userSender = new UserViewModel();
+                        //userSender.Id = senderProfile.Id;
+                        //userSender.Username = senderProfile.UserName;
+                        //userSender.Age = senderProfile.Age;
+                        //userSender.Description = senderProfile.Description;
+                        //userSender.Gender = senderProfile.Gender;
+                        friendRequests.Add(senderProfile);
+                        //friendRequests.Status = request.Status;
+                        enumerableFriendRequests = friendRequests;
                     }
                 }
-                return View(friendlist);
+                return View(enumerableFriendRequests);
             }
-            return View(friendlist);
+            TempData["noFriend"] = ViewData["noFriends"] = false;
+            TempData["noFriendRequest"] = ViewData["noFriendRequests"] = true;
+            return RedirectToAction("NoFriendOrFriendRequest");
         }
-        
+        public ActionResult NoFriendOrFriendRequest()
+        {
+            ViewBag.noFriends = TempData["noFriend"];
+            ViewBag.noFriendRequests = TempData["noFriendRequest"];
+
+            return View();
+        }
     }
 }
